+++ FaunaDB.Client/Types/NullV.cs
@@ -10,7 +10,7 @@ namespace FaunaDB.Types
     {
         public static readonly Value Instance = new NullV();
 
-        NullV() {}
+        internal NullV() {}
 
         protected internal override void WriteJson(JsonWriter writer)
         {